fix: crash because of attempting to rewrite existing file in snapshot
The `dir` DSL function copies an entire directory from the resource folder to the snapshot. If a
`file` DSL function call is nested inside a `dir` call, the snapshot crashes due to an already
existing file. This fix checks if the file already exists in the snapshot before attempting to make
a copy.

+++ core/src/main/kotlin/io/redgreen/fluid/snapshot/InMemorySnapshot.kt
@@ -128,7 +128,9 @@ class InMemorySnapshot private constructor(
     val source = if (resource.isSameAsDestination()) destination else resource.filePath
     createMissingDirectoriesInPath(destination)
     classLoader.getResourceAsStream(source)?.use { inputStream ->
-      Files.copy(inputStream, snapshotRoot.resolve(destination))
+      if (!Files.exists(snapshotRoot.resolve(destination))) {
+        Files.copy(inputStream, snapshotRoot.resolve(destination))
+      }
     } ?: throw IllegalStateException("Unable to find source file: '$source'") // TODO: Add tests for missing files and templates
   }
